wrap notify ws client
